Turning OpModeTwo
This set f code is intended to have strafing on right stick, and basic turning and forward and backward drive on the left stick. I also put in a perma-drive command on OpModeOne.

This also tracks the motor strength in OpModeTwo

After testing I decided to trun down the speeds as well